--- a/examples/webgl_mesh_batch.html
+++ b/examples/webgl_mesh_batch.html
@@ -302,6 +302,7 @@
 
 			}
 
+			// perform a fast-sort using the hybrid radix sort function
 			radixSort( list, options );
 
 		}