--- a/src/objects/BatchedMesh.js
+++ b/src/objects/BatchedMesh.js
@@ -130,6 +130,7 @@ class BatchedMesh extends Mesh {
 		this.sortObjects = true;
 		this.boundingBox = null;
 		this.boundingSphere = null;
+		this.customSort = null;
 
 		this._drawRanges = [];
 		this._reservedRanges = [];
@@ -298,6 +299,13 @@ class BatchedMesh extends Mesh {
 			return finalRange.indexStart + finalRange.indexCount;
 
 		}
+
+	}
+
+	setCustomSort( func ) {
+
+		this.customSort = func;
+		return this;
 
 	}
 
@@ -982,20 +990,22 @@ class BatchedMesh extends Mesh {
 
 			// Sort the draw ranges and prep for rendering
 			const list = _renderList.list;
-			console.time( 'SORT TIME' );
-			if ( window.RADIX === true ) {
+			const customSort = this.customSort;
+			if ( customSort === null ) {
 
-				this.__options = this.__options || {
-					get: el => el.z,
-					aux: new Array( list.length ),
-				};
-				this.__options.reversed = material.transparent;
-
-				radixSort( list, this.__options );
+				list.sort( material.transparent ? sortTransparent : sortOpaque );
 
 			} else {
 
-				list.sort( material.transparent ? sortTransparent : sortOpaque );
+				customSort( list, material.transparent );
+
+				// this.__options = this.__options || {
+				// 	get: el => el.z,
+				// 	aux: new Array( list.length ),
+				// };
+				// this.__options.reversed = material.transparent;
+
+				// radixSort( list, this.__options );
 
 			}