Related issue: https://github.com/mrdoob/three.js/pull/27202#issuecomment-1817640271

**Description**

For the sake of discussion here's a draft PR that shows the performance improvements using the suggested sort implementation from https://github.com/mrdoob/three.js/pull/27202#issuecomment-1817640271. If you want to try it you can run the batched example page and toggle the "useRadix" option to see the difference. The sort time is logged in the console.

**Performance**

Overall it looks like the new sort is providing roughly a 10x improvement in sort time over the native sort information. The drawback of the sort is that it requires more memory to store a temporary array:

| | hybrid sort | Array.sort |
|---|---|---|
| sort time | ~0.65ms | ~6.3ms |
| fps | ~94fps | ~74fps |

There seem to be occasional dips in fps with the hybrid sort to ~89fps on my machine but I'm not sure why.

Overall it's a meaningful improvement in complex cases. I'm not sure what opinions of others are on whether this should be included in three, though.

cc @Mugen87 @mrdoob @sciecode