@sciecode 2023-11-19
I get similar sorting times, but slightly better overall performance ( ~40% )

|   | hybrid sort | Array.sort |
| - | ----------- | ---------- |
| sort | ~0.69ms | ~6.5ms |
| fps | ~89fps | ~62fps |

I believe radix-sort can be a beneficial addition to three, but it's important to inform the users of the restrictions, as it is a less general function than `Array.sort`.

There are currently two important restrictions:
- It expects values in range [0, UINT32_MAX]
- `array.length` <= UINT32_MAX

The first restriction is not bad as it might seem, it does work on floating-point numbers if they are normalized to expected range, as Garrett did in the example. And the same thing can be done with signed 32-bit integers, like so:
```
// UINT32_MAX = 4_294_967_295
// INT32_MAX+1 = 2_147_483_648
const remap = (signed) => signed + INT32_MAX + 1;
```