@Mugen87 2023-11-19
TBH, I would prefer to keep this outside of the core since `BatchedMesh` is already a quite complex component and we should not over-engineer it. 

Maybe you could add something like `BatchedMesh.setCustomSort()` so every developer cann add a custom sort implementation if performance would be an issue.