@gkjohnson 2023-11-20
> TBH, I would prefer to keep this outside of the core since BatchedMesh is already a quite complex component and we should not over-engineer it.
>
> Could you add something like BatchedMesh.setCustomSort() so every developer can add a custom sort implementation if performance would be an issue?

It's a pretty significant improvement but admittedly likely only relevant for particularly complex cases. How would you feel about something like a `fastBatchedMeshSort` and / or `hybridRadixSort` to the three.js examples so this code is available?