@Mugen87 2023-11-20
So you mean moving `src/objects/radixSort.js` to the addons and make a new example that shows how to use it via `setCustomSort()`? That sounds good to me 👍 .